ontosearch/ext_search/requirements.txt
Flask==1.0.2
python-dotenv==0.10.1
rdflib==4.2.2
rdflib-jsonld==0.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
numpy==1.16.2
scipy==1.2.1
scikit-learn==0.20.3
pandas==0.24.1
nltk==3.4
pymongo==3.7.2
requests==2.21.0
requests-html==0.10.0
clint==0.5.1
beautifulsoup4==4.7.1
html5lib==1.0.1
tabulate==0.8.3
PyYAML==3.13
## The following requirements were added by pip freeze:
appdirs==1.4.3
args==0.1.0
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cssselect==1.0.3
fake-useragent==0.1.11
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.3
MarkupSafe==1.1.1
parse==1.12.0
pyee==6.0.0
pyparsing==2.3.1
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.8.0
pytz==2018.9
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
tqdm==4.31.1
urllib3==1.24.1
w3lib==1.20.0
webencodings==0.5.1
websockets==7.0
Werkzeug==0.14.1
WTForms==2.2.1
